refactor: moved collection schema into 'src'
